Worked on seller collection listing
